+++ examples/what.rs
@@ -2,10 +2,8 @@ use std::env;
 
 extern crate imghdr;
 
-
 fn main() {
-
     for path in env::args().skip(1) {
-        println!("{:?}", imghdr::open(&path));
+        println!("{:?}", imghdr::from_file(&path));
     }
 }